+++ R/cohens_d.R
@@ -24,9 +24,10 @@
 #' @param iv Character indicating the name of the column in \code{data} for the
 #'   independent variable
 #' @param formula A formula of the form \code{lhs ~ rhs} where \code{lhs} is a
-#'   numeric variable giving the data values (dependent variable) and \code{rhs}
-#'   a factor with two levels giving the corresponding groups (independent
-#'   variable).
+#'   numeric variable giving the data values and \code{rhs}
+#'   either \code{1} for one sample or paired data or a factor with two levels
+#'   giving the corresponding groups. If \code{lhs} is of class \code{"Pair"}
+#'   and \code{rhs} is \code{1}, Cohen's d for paired data will be calculated.
 #' @param ttest An object of class \code{htest} (a call to either \code{t_test}
 #'   (preferred) or \code{t.test}).
 #' @param ... Further arguments passed to methods.
@@ -42,10 +43,11 @@
 #' # or
 #' cohens_d(sleep, dv = "extra", iv = "group", paired = TRUE)
 #' # formula interface
-#' cohens_d(extra ~ group, sleep, paired = TRUE)
+#' sleep2 <- reshape(sleep, direction = "wide", idvar = "ID", timevar = "group")
+#' cohens_d(Pair(extra.1, extra.2) ~ 1, sleep2, paired = TRUE)
 #'
 #' # Or pass a call to t_test or t.test
-#' cohens_d(t_test(extra ~ group, sleep, paired = TRUE))
+#' cohens_d(t_test(Pair(extra.1, extra.2) ~ 1, sleep2))
 #' @export
 cohens_d <- function(...) UseMethod("cohens_d")
 
@@ -108,16 +110,16 @@ cohens_d.data.frame <- function(data, dv, iv, paired = FALSE,
 }
 
 #' @rdname cohens_d
-#' @importFrom stats model.frame setNames
 #' @export
-cohens_d.formula <- function(formula, data, paired = FALSE,
+cohens_d.formula <- function(formula, data,
                              corr = c("none", "hedges_g", "glass_delta"),
                              na.rm = FALSE, ...)
 {
   corr <- match.arg(corr)
 
-  mf <- model.frame(formula, data)
-  .data <- setNames(split(mf[[1]], mf[[2]]), c("x", "y"))
+  .data <- extract_data_formula(formula, data, ...)
+
+  paired <- grepl("Pair\\(*., *.\\)", as.character(formula)[2])
 
   do.call("cohens_d", c(.data, paired = paired, corr = corr, na.rm = na.rm))
 }