+++ R/t_test.R
@@ -53,14 +53,13 @@ t_test.default <- function(x, y = NULL,
 }
 
 #' @rdname t_test
-#' @importFrom stats model.frame na.omit setNames t.test
+#' @importFrom stats t.test
 #' @export
 t_test.formula <- function(formula, data, subset, na.action, ...)
 {
   t <- t.test(formula = formula, data = data, ...)
 
-  mf <- na.omit(model.frame(formula, data, drop.unused.levels = TRUE))
-  t$data <- setNames(split(mf[[1]], mf[[2]]), c("x", "y"))
+  t[["data"]] <- extract_data_formula(formula, data, ...)
 
   t
 }